minikube-project/pyproject.toml
[project]
name = "user"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faust>=1.4.9",
    "influxdb-client>=1.49.0",
    "kafka-python>=2.2.15",
    "openpyxl>=3.1.5",
    "paho-mqtt>=2.1.0",
    "pandas>=2.3.3",
]
